pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "grbl2image_AlanFromJapan"
version = "1.0.0"
authors = [
  { name="AlanFromJapan", email="alanfromjapan@gmail.com" },
]
description = "Generates an image (PNG, JPEG, etc.) from a GRBL file (.NC, .GC, ...), in Python using PILlow (python3)"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "Pillow>=9.0"
]

[project.urls]
"Homepage" = "https://github.com/AlanFromJapan/grbl2image"
"Bug Tracker" = "https://github.com/AlanFromJapan/grbl2image/issues"